Cleanup of char_xml
mainly PEP8 and documentation cleanup including a bugfix in the event handling for weapons.

char_xml now has no skill_xml in it for new skills but has to be passed skill elements.

changes in itemeditor, skillselector and traitselector implement minor changes in the
return values of the char_xml